requirements.txt
Flask==3.0.2
Jinja2==3.1.3
keras==2.15.0
numpy==1.26.4
pillow==10.2.0
tensorflow==2.15.0
